Fix security bug SER-493, add more permission check in user-group-membership deletion
Summary:
1. Fix security bug SER-493, add more permission check in user-group-membership deletion
2. remove the pii.update and pii.delete from the identity.pii scope, this should be service rights only.

Test Plan: supercycle

Reviewers: liangfuxia, kg, shuz

Reviewed By: shuz

Subscribers: all

Differential Revision: https://code.silkcloud.info/D1080

+++ identity/identity-rest/src/main/groovy/com/junbo/identity/rest/resource/v1/UserGroupMembershipResourceImpl.groovy
@@ -157,8 +157,15 @@ class UserGroupMembershipResourceImpl implements UserGroupMembershipResource {
 
     @Override
     Promise<Void> delete(UserGroupId userGroupId) {
-        return userGroupValidator.validateForGet(userGroupId).then {
-            return userGroupRepository.delete(userGroupId)
+        return userGroupValidator.validateForGet(userGroupId).then { UserGroup userGroup ->
+            def callback = authorizeCallbackFactory.create(userGroup.groupId)
+            return RightsScope.with(authorizeService.authorize(callback)) {
+                if (!AuthorizeContext.hasRights('delete') && AuthorizeContext.currentUserId != userGroup.userId) {
+                    throw AppCommonErrors.INSTANCE.forbidden().exception()
+                }
+                return userGroupRepository.delete(userGroupId)
+            }
+
         }
     }
